Update join to use response DTO
