+++ src/test/java/teammates/test/cases/webapi/GetCourseJoinStatusActionTest.java
@@ -6,8 +6,8 @@ import org.testng.annotations.Test;
 import teammates.common.util.Const;
 import teammates.common.util.StringHelper;
 import teammates.ui.webapi.action.GetCourseJoinStatusAction;
-import teammates.ui.webapi.action.GetCourseJoinStatusAction.JoinStatus;
 import teammates.ui.webapi.action.JsonResult;
+import teammates.ui.webapi.output.JoinStatus;
 
 /**
  * SUT: {@link GetCourseJoinStatusAction}.
@@ -56,7 +56,7 @@ public class GetCourseJoinStatusActionTest extends BaseActionTest<GetCourseJoinS
         assertEquals(HttpStatus.SC_OK, r.getStatusCode());
 
         JoinStatus output = (JoinStatus) r.getOutput();
-        assertTrue(output.isHasJoined());
+        assertTrue(output.getHasJoined());
         assertNull(output.getUserId());
 
         ______TS("Normal case: student is not registered");
@@ -75,7 +75,7 @@ public class GetCourseJoinStatusActionTest extends BaseActionTest<GetCourseJoinS
         assertEquals(HttpStatus.SC_OK, r.getStatusCode());
 
         output = (JoinStatus) r.getOutput();
-        assertFalse(output.isHasJoined());
+        assertFalse(output.getHasJoined());
         assertEquals("unreg.user", output.getUserId());
 
         ______TS("Failure case: regkey is not valid for student");
@@ -106,7 +106,7 @@ public class GetCourseJoinStatusActionTest extends BaseActionTest<GetCourseJoinS
         assertEquals(HttpStatus.SC_OK, r.getStatusCode());
 
         output = (JoinStatus) r.getOutput();
-        assertTrue(output.isHasJoined());
+        assertTrue(output.getHasJoined());
         assertNull(output.getUserId());
 
         ______TS("Normal case: instructor is not registered");
@@ -125,7 +125,7 @@ public class GetCourseJoinStatusActionTest extends BaseActionTest<GetCourseJoinS
         assertEquals(HttpStatus.SC_OK, r.getStatusCode());
 
         output = (JoinStatus) r.getOutput();
-        assertFalse(output.isHasJoined());
+        assertFalse(output.getHasJoined());
         assertEquals("unreg.user", output.getUserId());
 
         ______TS("Failure case: regkey is not valid for instructor");
